+++ R/fortify-igraph.R
@@ -22,35 +22,33 @@
 fortify.igraph <- function(model, data = NULL, layout = igraph::nicely(),
                            arrow.gap = ifelse(igraph::is.directed(x), 0.025, 0),
                            by = NULL, ...) {
-  x <- model
-
   # node placement
   if (
     class(layout) == "matrix" &&
-      nrow(layout) == igraph::gorder(x) &&
+      nrow(layout) == igraph::gorder(model) &&
       ncol(layout) == 2
   ) {
     nodes <- layout[, 1:2 ]
   } else {
-    nodes <- igraph::layout_(x, layout, ...)
+    nodes <- igraph::layout_(model, layout, ...)
   }
 
   # store coordinates
   nodes <- data.frame(nodes)
-  colnames(nodes) <- c("x", "y")
+  names(nodes) <- c("x", "y")
 
   # rescale coordinates
   nodes$x <- scale_safely(nodes$x)
   nodes$y <- scale_safely(nodes$y)
 
   # import vertex attributes
-  if (length(igraph::list.vertex.attributes(x))) {
+  if (length(igraph::list.vertex.attributes(model))) {
     nodes <- cbind(
       nodes,
       sapply(
-        igraph::list.vertex.attributes(x),
+        igraph::list.vertex.attributes(model),
         FUN = igraph::get.vertex.attribute,
-        graph = x,
+        graph = model,
         USE.NAMES = TRUE
       )
     )
@@ -58,7 +56,7 @@ fortify.igraph <- function(model, data = NULL, layout = igraph::nicely(),
 
 
   # edge list
-  edges <- igraph::as_edgelist(x, names = FALSE)
+  edges <- igraph::as_edgelist(model, names = FALSE)
 
   # edge list (if there are duplicated rows)
   if (nrow(edges[, 1:2]) > nrow(unique(edges[, 1:2]))) {
@@ -73,18 +71,25 @@ fortify.igraph <- function(model, data = NULL, layout = igraph::nicely(),
     x.length <- edges$xend - edges$x
     y.length <- edges$yend - edges$y
     arrow.gap <- edges$arrow.gap / sqrt(x.length^2 + y.length^2)
-    edges$xend <- edges$x + (1 - arrow.gap) * x.length
-    edges$yend <- edges$y + (1 - arrow.gap) * y.length
+    # edges$xend <- edges$x + (1 - arrow.gap) * x.length
+    # edges$yend <- edges$y + (1 - arrow.gap) * y.length
+    edges <- transform(
+      edges,
+      # x = x + arrow.gap * x.length,
+      # y = y + arrow.gap * y.length,
+      xend = x + (1 - arrow.gap) * x.length,
+      yend = y + (1 - arrow.gap) * y.length
+    )
   }
 
   # import edge attributes
-  if (length(igraph::list.edge.attributes(x))) {
+  if (length(igraph::list.edge.attributes(model))) {
     edges <- cbind(
       edges,
       sapply(
-        igraph::list.edge.attributes(x),
+        igraph::list.edge.attributes(model),
         FUN = igraph::get.edge.attribute,
-        graph = x,
+        graph = model,
         USE.NAMES = TRUE
       )
     )
@@ -104,16 +109,16 @@ fortify.igraph <- function(model, data = NULL, layout = igraph::nicely(),
 
   # panelize nodes (for temporal networks)
   if (!is.null(by)) {
-    nodes <- lapply(sort(unique(edges[, by ])), function(x) {
+    nodes <- lapply(sort(unique(edges[, by])), function(x) {
       y <- nodes
-      y[, by ] <- x
+      y[, by] <- x
       y
     })
     nodes <- do.call(rbind, nodes)
   }
 
-  # return a data frame with network.size(x) + network.edgecount(x) rows,
-  # or length(unique(edges[, by ])) * network.size(x) + network.edgecount(x)
+  # return a data frame with network.size(model) + network.edgecount(model) rows,
+  # or length(unique(edges[, by])) * network.size(model) + network.edgecount(model)
   # rows if the nodes have been panelized
-  unique(rbind(nodes, edges[ !is.na(edges$xend), ]))
+  unique(rbind(nodes, edges[!is.na(edges$xend), ]))
 }